Fix implicit cast warnings
These are found by some internal analysis scripts, so I just helped to fix them here to silent the warnings.

+++ Classes/Editing/ArgumentInputViews/FLEXArgumentInputTextView.m
@@ -128,7 +128,7 @@
 }
 
 - (CGFloat)inputTextViewHeight {
-    return ceil([[self class] inputFont].lineHeight * self.numberOfInputLines) + 16.0;
+    return ceil([[self class] inputFont].lineHeight * (CGFloat)self.numberOfInputLines) + 16.0;
 }
 
 - (CGSize)sizeThatFits:(CGSize)size {
